+++ Source/Tests/Gapotchenko.FX.Math.Combinatorics.Tests/PermutationTests.cs
@@ -14,7 +14,7 @@ public class PermutationTests
         int cardinality = Permutations.Cardinality(source.Length);
         Assert.AreEqual(2, cardinality);
 
-        var p = Permutations.Of(source).ReifyList();
+        var p = Permutations.Of(source).ReifyCollection();
         Assert.AreEqual(cardinality, p.Count);
 
         var s = p
@@ -35,7 +35,7 @@ public class PermutationTests
         int cardinality = Permutations.Cardinality(source.Length);
         Assert.AreEqual(2, cardinality);
 
-        var p = Permutations.Of(source).Distinct().ReifyList();
+        var p = Permutations.Of(source).Distinct().ReifyCollection();
         Assert.AreEqual(cardinality, p.Count);
 
         var s = p
@@ -69,7 +69,7 @@ public class PermutationTests
         int cardinality = Permutations.Cardinality(source.Length);
         Assert.AreEqual(2, cardinality);
 
-        var p = Permutations.Of(source).ReifyList();
+        var p = Permutations.Of(source).ReifyCollection();
         Assert.AreEqual(cardinality, p.Count);
 
         var s = p
@@ -91,7 +91,7 @@ public class PermutationTests
         int cardinality = Permutations.Cardinality(source.Length) / countOfDuplicateElements;
         Assert.AreEqual(1, cardinality);
 
-        var p = Permutations.Of(source).Distinct().ReifyList();
+        var p = Permutations.Of(source).Distinct().ReifyCollection();
         Assert.AreEqual(cardinality, p.Count);
 
         var s = p
@@ -128,7 +128,7 @@ public class PermutationTests
         int cardinality = Permutations.Cardinality(source.Length);
         Assert.AreEqual(6, cardinality);
 
-        var p = Permutations.Of(source).ReifyList();
+        var p = Permutations.Of(source).ReifyCollection();
         Assert.AreEqual(cardinality, p.Count);
 
         var s = p
@@ -157,7 +157,7 @@ public class PermutationTests
         int cardinality = Permutations.Cardinality(source.Length);
         Assert.AreEqual(6, cardinality);
 
-        var p = Permutations.Of(source).Distinct().ReifyList();
+        var p = Permutations.Of(source).Distinct().ReifyCollection();
         Assert.AreEqual(cardinality, p.Count);
 
         var s = p
@@ -227,7 +227,7 @@ public class PermutationTests
         int cardinality = Permutations.Cardinality(source.Length) / countOfDuplicateElements;
         Assert.AreEqual(3, cardinality);
 
-        var p = Permutations.Of(source).Distinct().ReifyList();
+        var p = Permutations.Of(source).Distinct().ReifyCollection();
         Assert.AreEqual(cardinality, p.Count);
 
         var s = p
@@ -267,7 +267,7 @@ public class PermutationTests
         int cardinality = Permutations.Cardinality(source.Length) / countOfDuplicateElements;
         Assert.AreEqual(3, cardinality);
 
-        var p = Enumerable.Distinct(Permutations.Of(source)).ReifyList();
+        var p = Enumerable.Distinct(Permutations.Of(source)).ReifyCollection();
         Assert.AreEqual(cardinality, p.Count);
 
         var s = p